requirements.txt
streamlit
sentence-transformers
numpy
scikit-learn
google-genai
PyPDF2
python-dotenv
